Enhance snacks UI
